requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.30.0
openai==1.10.0
neo4j==5.16.0
chromadb==0.4.22
sentence-transformers==2.3.1
biopython==1.83
python-dotenv==1.0.0
tavily-python==0.3.0
torch==2.1.2
pydantic==2.5.3